applied MVVM architecture / some compile errors
