+++ app/src/main/res/layout/recorder_fragment.xml
@@ -4,6 +4,12 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:android="http://schemas.android.com/apk/res/android">
 
+    <data>
+        <variable
+            name="recorderViewModel"
+            type="berlin.eloquent.eloquentandroid.recorder.RecorderViewModel" />
+    </data>
+
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent">
@@ -17,7 +23,7 @@
             android:layout_marginLeft="32dp"
             android:layout_marginBottom="200dp"
             android:background="@drawable/custom_border"
-
+            android:onClick="@{() -> recorderViewModel.startRecording()}"
             android:padding="8dp"
             android:src="@drawable/ic_mic"
             app:layout_constraintBottom_toBottomOf="parent"
@@ -36,6 +42,7 @@
             android:background="@drawable/custom_border"
             android:padding="8dp"
             android:src="@drawable/ic_pause"
+            android:onClick="@{() -> recorderViewModel.pauseRecording()}"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent" />
@@ -51,6 +58,7 @@
             android:background="@drawable/custom_border"
             android:padding="8dp"
             android:src="@drawable/ic_stop"
+            android:onClick="@{() -> recorderViewModel.stopRecording()}"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent" />
 
@@ -67,6 +75,7 @@
             android:background="@drawable/custom_border"
             android:padding="8dp"
             android:src="@drawable/ic_play_arrow"
+            android:onClick="@{() -> recorderViewModel.playRecording()}"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent" />
